refactor(gcp_link_account): updated rename account mutation with update account mutation
